feat: integrate api gateway with auth-service

+++ AuthService/src/main/java/com/javagrind/authservice/services/impl/AuthServiceImpl.java
@@ -4,19 +4,19 @@ import com.javagrind.authservice.dto.request.Auth.LoginRequest;
 import com.javagrind.authservice.dto.request.Auth.LogoutRequest;
 import com.javagrind.authservice.dto.response.LoginResponse;
 import com.javagrind.authservice.dto.response.LogoutResponse;
-import com.javagrind.authservice.repositories.RoleRepository;
-import com.javagrind.authservice.repositories.UserRepository;
+import com.javagrind.authservice.dto.response.TokenResponse;
 import com.javagrind.authservice.security.jwt.JwtUtils;
 import com.javagrind.authservice.security.services.RedisService;
 import com.javagrind.authservice.security.services.UserDetailsImpl;
 import com.javagrind.authservice.services.AuthService;
+import io.jsonwebtoken.JwtException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import java.sql.Timestamp;
 
 @Service
 @RequiredArgsConstructor
@@ -47,11 +47,16 @@ public class AuthServiceImpl implements AuthService {
 
     @Override
     public Object logout(LogoutRequest request){
-
-        if(redisService.isThere(request.getEmail())!=null){
-            redisService.destroyToken(request.getEmail());
-        }
+        if(redisService.isThere(request.getEmail())!=null)  {redisService.destroyToken(request.getEmail());}
 
         return new LogoutResponse(request.getEmail());
+    }
+
+    @Override
+    public Object validate(String jwt){
+        if (jwt != null && jwtUtils.validateJwtToken(jwt) && redisService.isThere((String)jwtUtils.getUserNameFromJwtToken(jwt) )!=null)
+            return new TokenResponse(jwt,jwtUtils.getUserNameFromJwtToken(jwt),new Timestamp(jwtUtils.getExpiredAt(jwt)));
+        else
+            throw new JwtException("JWT not valid");
     }
 }
